Rename file to be more sane, add 'chat' badge
